Fixes MAISTRA-2324:
port of tls: separate out cert validation logic from ContextImpl (#14757)
